requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysql-connector==2.2.9
prompt-toolkit==3.0.5
requests==2.23.0
urllib3==1.25.8
wcwidth==0.1.9